+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -4099,7 +4099,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01KG2RZQB6NY02X6PDEWS9295S"),
+  id: z.string().default("01KG4856ZQT5BFZE3KVBEJQWTP"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() =>