+++ packages/orq-rc/RELEASES.md
@@ -6229,3 +6229,13 @@ Based on:
 - [typescript v4.3.0-rc.7] packages/orq-rc
 ### Releases
 - [NPM v4.3.0-rc.7] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.7 - packages/orq-rc
+
+## 2026-01-29 06:51:10
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.700.2 (2.801.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.3.0-rc.8] packages/orq-rc
+### Releases
+- [NPM v4.3.0-rc.8] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.8 - packages/orq-rc